// carry-app/src/dts/color.d.ts
interface Color {
  id: string
  name: string
  background: string
  lavender: string
  middle: string
  contrast: string
  text: string
  shadowColor: string

  green: string
  green2: string
  blue: string
  blue0: string
  blue1: string
  blue2: string
  blue3: string
  red: string
  yellow: string
  orange: string
  orange2: string
  purple: string
  purple1: string
  purple2: string
  notification: string

  gray: string
  gray2: string
  gray3: string
  gray4: string
  gray5: string
  gray6: string
  gray7: string
  gray8: string
  gray9: string
  gray10: string
  white: string
  barStyle: 'dark-content' | 'light-content' | 'default'
  black: string
  black2: string
  chat?: any
  jesusWords: string
  whiteSmoke: string

  primary: string
  heading: string
  bodyText: string
  accent: string
  accent2: string
  lightBackground: string
  lightText: string
  lightTextSecondary: string
  transparent: string
  lightAccent: string
}

export default Color